Незначительные доделки связанные с валютой сайта...

+++ apps/cart/models.py
@@ -132,8 +132,9 @@ class Order(models.Model):
     @property
     def order_sum(self, ):
         all_products_sum = 0
+        from decimal import Decimal
         for product in self.products:
-            all_products_sum += int(product.summ_of_quantity().replace('.', ',', ), )
+            all_products_sum += Decimal(product.summ_of_quantity().replace('.', ',', ), )
         return all_products_sum
 
     def __unicode__(self):
